Move range conversions up a module
